Review controller: Split up controller into three files with individual responsibilities
